Update error handling for route
