+++ packages/widget-v2/src/pages/SwapPage/SwapDetailModal.tsx
@@ -5,7 +5,8 @@ import { SmallText } from "@/components/Typography";
 import { RouteArrow } from "@/icons/RouteArrow";
 import { SwapPageFooterItems } from "./SwapPageFooter";
 import { useAtom, useAtomValue } from "jotai";
-import { skipChainsAtom, skipRouteAtom } from "@/state/skipClient";
+import { skipChainsAtom } from "@/state/skipClient";
+import { skipRouteAtom } from "@/state/route";
 import { useMemo, useState } from "react";
 import { swapSettingsAtom } from "@/state/swapPage";
 import { formatUSD } from "@/utils/intl";
