+++ packages/widget-v2/src/pages/SwapPage/SwapPageFooter.tsx
@@ -5,7 +5,7 @@ import { SkipLogoIcon } from "@/icons/SkipLogoIcon";
 import { SpeedometerIcon } from "@/icons/SpeedometerIcon";
 import { useAtomValue } from "jotai";
 import { convertSecondsToMinutesOrHours } from "@/utils/number";
-import { skipRouteAtom } from "@/state/skipClient";
+import { skipRouteAtom } from "@/state/route";
 import { SignatureIcon } from "@/icons/SignatureIcon";
 import pluralize from "pluralize";
 import { styled } from "styled-components";
